requirements.txt
django==2.2.4
psycopg2-binary==2.7.7
celery==4.3.0
requests==2.21.0
beautifulsoup4==4.7.1
django-elasticsearch-dsl==7.0.0
flower==0.9.3
django_celery_beat==1.4.0
django-redis==4.10.0
lxml==4.3.3
defusedxml==0.6.0
django-celery-results==1.0.4
django-simple-history==2.7.3
pyyaml==5.1.2
mysqlclient==1.4.4
gunicorn==19.9.0
